3.12* moving Node class
