+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -1,5 +1,38 @@
 package edu.kis.vh.nursery.list;
 
+
+ class Node {
+
+    public int getValue() {
+        return value;
+    }
+
+
+    private int value;
+    private Node prev, next;
+
+    public Node(int i) {
+        value = i;
+    }
+
+    public Node getPrev() {
+        return prev;
+    }
+
+    public void setPrev(Node prev) {
+        this.prev = prev;
+    }
+
+    public Node getNext() {
+        return next;
+    }
+
+    public void setNext(Node next) {
+        this.next = next;
+    }
+
+}
+
 public class IntLinkedList {
 
     private Node last;
